Initial configuration, profiles and remotes added
This commit also adds a remote artifactory server which is
currently a private personal server and just for tryouts.
This needs to change in due time to an UM managed server.
Either in the cloud or on a VM.

There are three profiles added to the configuration. All
of them depend on the default autodetected profile.
This default profile can be created using the command
`conan profile new --detect default`

The configuration as whole can be installed with the command:
`conan config install https://github.com/ultimaker/conan-config.git`

Contributes to CURA-8785