Generate and deploy files for a curapackage
It will copy the files from the dependencies in the deps
folder. It will make this folder a Python module with an
__init__.py which should set the paths to the correct
locations such that Cura (or Python) can find the files

If you want to create a curapackage supporting multiple
python_versions you would have to run the `conan install -o python_version` command multiple times.

Contributes to CURA-9068