Make the VirtulPythonEnv more robust
It should now be able to also create a venv with another interpreter.
Provided it is installed next to the current interpreter on the same
base Path.

Trying to make execution of the run commands robust on Windows, till no
avail. It states:

```powershell
----Running------
> C:\'Program Files'\Python310\python.exe -m venv venv
-----------------
'C:\'Program' is not recognized as an internal or external command,
operable program or batch file.
```

But if I run the ` C:\'Program Files'\Python310\python.exe -m venv venv`
in the same shell manually it executes correctly.

Contributes to CURA-9365