Add cura_wasm.jinja profile
The new cura_wasm.jinja configuration profile has been added. This profile includes settings and options specific to Emscripten and WebAssembly (wasm), along with certain build, tool, and requirement configurations like skipping tests and disabling plugins.

This profile was already tested during the work on NP-5 and is merged on the dev branch

Contribute to NP202